Renamed dwcMapping.py to mapping.py Updated references and cleaned imports of affected files Removed redundant retrieveMapping.py
