--- src/tools/generic/retrieveMapping.py
@@ -1,6 +1,0 @@
-from lib.processing.dwcMapping import MapRetriever
-
-if __name__ == "__main__":
-    retriever = MapRetriever()
-    retriever.run()
-
